refactor: organize frontend by entity
