+++ src/renderer/entities/CrudConfig.ts
@@ -1,0 +1,7 @@
+export interface CrudConfig {
+  entity: string;
+  title: string;
+  columns: (string | { titulo: string; clave: string })[];
+  searchFields: string[];
+  formInputs: any[];
+}
