+++ src/renderer/entities/clientes/Page.tsx
@@ -1,0 +1,6 @@
+import CrudPage from '../../pages/CrudPage';
+import config from './config';
+
+export default function PageClientes() {
+  return <CrudPage config={config} />;
+}
