+++ src/renderer/entities/item-venta/Page.tsx
@@ -1,0 +1,6 @@
+import CrudPage from '../../pages/CrudPage';
+import config from './config';
+
+export default function PageItemVenta() {
+  return <CrudPage config={config} />;
+}
