+++ src/renderer/entities/metodo-pago/Page.tsx
@@ -1,0 +1,6 @@
+import CrudPage from '../../pages/CrudPage';
+import config from './config';
+
+export default function PageMetodoPago() {
+  return <CrudPage config={config} />;
+}
