+++ src/renderer/entities/productos/Page.tsx
@@ -1,0 +1,6 @@
+import CrudPage from '../../pages/CrudPage';
+import config from './config';
+
+export default function PageProductos() {
+  return <CrudPage config={config} />;
+}
