--- src/renderer/pages/PageMetodoPago.tsx
@@ -1,16 +1,0 @@
-import CrudPage from './CrudPage';
-
-export default function PageMetodoPago() {
-  const config = {
-    entity: 'metodo-pago',
-    title: 'Métodos de Pago',
-    columns: ['id', 'nombre', 'descripcion'],
-    searchFields: ['nombre', 'descripcion'],
-    formInputs: [
-      { name: 'nombre', label: 'Nombre', type: 'text', required: true },
-      { name: 'descripcion', label: 'Descripción', type: 'text' },
-    ],
-  };
-
-  return <CrudPage config={config} />;
-}
